requirements.txt
flask==3.0.0
gunicorn==21.2.0
ccxt==4.2.0
pandas==2.1.4
python-dotenv==1.0.0
numpy==1.26.3
